possibility to use Rails routes for Devise.cas_action_url
refactoring `Devise.cas_action_url` method into a class.

The first time it's needed it will create a new class with
Rails routes helpers included (also engines).

It will try to use the routes if it can, looking it up in a similar way that Devise does.

This allow to use devise_cas_authenticatable with a Devise inside of an engine.